requirements.txt
Django==3.2.6
djangorestframework==3.12.4
celery==5.1.2
redis==3.5.3
django-celery-beat==2.2.1
psycopg2==2.9.1
requests==2.26.0
beautifulsoup4==4.9.3
lxml==4.6.3